include financial benefit calculation

+++ types/financials.jl
@@ -1,6 +1,6 @@
 mutable struct Financial
     inflation::Float64
     apr::Float64
-    term::Float64
+    term::Int64
     XhR::Float64
 end
